split the context code
