+++ front/components/data_source_view/context/types.ts
@@ -1,0 +1,16 @@
+import type {
+  DataSourceViewCategoryWithoutApps,
+  DataSourceViewContentNode,
+  SpaceType,
+} from "@app/types";
+
+export type DataSourceBuilderTree = {
+  in: string[];
+  notIn: string[];
+};
+
+export type NavigationHistoryEntryType =
+  | { type: "root" }
+  | { type: "space"; space: SpaceType }
+  | { type: "category"; category: DataSourceViewCategoryWithoutApps }
+  | { type: "node"; node: DataSourceViewContentNode };
